taking a clean pass on indexing
