+++ crates/runmat-vm/src/indexing/mod.rs
@@ -1,4 +1,5 @@
 pub mod end_expr;
+pub mod plan;
 pub mod read_linear;
 pub mod read_slice;
 pub mod selectors;
