+++ crates/runmat-vm/src/indexing/read_slice.rs
@@ -1,11 +1,9 @@
-use crate::bytecode::EndExpr;
+use crate::indexing::plan::{build_index_plan, IndexPlan};
 use crate::indexing::selectors::{
-    build_slice_plan, build_slice_selectors, index_scalar_from_value, materialize_index_value,
-    SlicePlan, SliceSelector,
+    build_slice_selectors, index_scalar_from_value, materialize_index_value, SliceSelector,
 };
 use runmat_builtins::{CellArray, ComplexTensor, StringArray, Tensor, Value};
 use runmat_runtime::RuntimeError;
-use std::future::Future;
 
 pub fn build_numeric_subsref_cell(numeric: &[Value]) -> Result<Value, RuntimeError> {
     let cell = CellArray::new(numeric.to_vec(), 1, numeric.len())
@@ -245,7 +243,7 @@ pub async fn read_tensor_slice_nd(
     if let Some(value) = try_tensor_slice_2d_fast_path(tensor, dims, &selectors)? {
         return Ok(value);
     }
-    let plan = build_slice_plan(&selectors, dims, &tensor.shape)?;
+    let plan = build_index_plan(&selectors, dims, &tensor.shape)?;
     if plan.indices.is_empty() {
         let out_tensor =
             Tensor::new(Vec::new(), plan.output_shape).map_err(|e| format!("Slice error: {e}"))?;
@@ -266,7 +264,7 @@ pub async fn read_tensor_slice_nd(
 
 pub fn read_tensor_slice_from_plan(
     tensor: &Tensor,
-    plan: &SlicePlan,
+    plan: &IndexPlan,
 ) -> Result<Value, RuntimeError> {
     if plan.indices.is_empty() {
         let out_tensor = Tensor::new(Vec::new(), plan.output_shape.clone())
@@ -295,13 +293,13 @@ pub async fn read_complex_slice(
 ) -> Result<Value, RuntimeError> {
     let selectors =
         build_slice_selectors(dims, colon_mask, end_mask, numeric, &tensor.shape).await?;
-    let plan = build_slice_plan(&selectors, dims, &tensor.shape)?;
+    let plan = build_index_plan(&selectors, dims, &tensor.shape)?;
     read_complex_slice_from_plan(tensor, &plan)
 }
 
 pub fn read_complex_slice_from_plan(
     tensor: &ComplexTensor,
-    plan: &SlicePlan,
+    plan: &IndexPlan,
 ) -> Result<Value, RuntimeError> {
     if plan.indices.is_empty() {
         let empty = ComplexTensor::new(Vec::new(), plan.output_shape.clone())
@@ -343,13 +341,13 @@ pub async fn read_gpu_slice(
 ) -> Result<Value, RuntimeError> {
     let base_shape = handle.shape.clone();
     let selectors = build_slice_selectors(dims, colon_mask, end_mask, numeric, &base_shape).await?;
-    let plan = build_slice_plan(&selectors, dims, &base_shape)?;
+    let plan = build_index_plan(&selectors, dims, &base_shape)?;
     read_gpu_slice_from_plan(handle, &plan)
 }
 
 pub fn read_gpu_slice_from_plan(
     handle: &runmat_accelerate_api::GpuTensorHandle,
-    plan: &SlicePlan,
+    plan: &IndexPlan,
 ) -> Result<Value, RuntimeError> {
     let provider = runmat_accelerate_api::provider().ok_or_else(|| {
         crate::interpreter::errors::mex(
@@ -606,7 +604,7 @@ pub async fn read_string_slice(
     }
 }
 
-pub fn gather_string_slice(sa: &StringArray, plan: &SlicePlan) -> Result<Value, RuntimeError> {
+pub fn gather_string_slice(sa: &StringArray, plan: &IndexPlan) -> Result<Value, RuntimeError> {
     if plan.indices.is_empty() {
         let empty = StringArray::new(Vec::new(), plan.output_shape.clone())
             .map_err(|e| format!("Slice error: {e}"))?;
@@ -636,306 +634,4 @@ pub fn gather_string_slice(sa: &StringArray, plan: &SlicePlan) -> Result<Value, 
     let out_sa = StringArray::new(out, plan.output_shape.clone())
         .map_err(|e| format!("Slice error: {e}"))?;
     Ok(Value::StringArray(out_sa))
-}
-
-fn total_len_from_shape(shape: &[usize]) -> usize {
-    if shape.is_empty() {
-        1
-    } else {
-        shape.iter().product()
-    }
-}
-
-#[derive(Clone)]
-enum ExprSel {
-    Colon,
-    Scalar(usize),
-    Indices(Vec<usize>),
-    Range {
-        start: i64,
-        step: i64,
-        end_off: EndExpr,
-    },
-}
-
-pub struct ExprGatherSpec<'a> {
-    pub dims: usize,
-    pub colon_mask: u32,
-    pub end_mask: u32,
-    pub range_dims: &'a [usize],
-    pub range_params: &'a [(f64, f64)],
-    pub range_start_exprs: &'a [Option<EndExpr>],
-    pub range_step_exprs: &'a [Option<EndExpr>],
-    pub range_end_exprs: &'a [EndExpr],
-    pub numeric: &'a [Value],
-    pub shape: &'a [usize],
-}
-
-pub async fn build_expr_gather_plan<ResolveEnd, Fut>(
-    spec: ExprGatherSpec<'_>,
-    mut resolve_end: ResolveEnd,
-) -> Result<SlicePlan, RuntimeError>
-where
-    ResolveEnd: FnMut(usize, &EndExpr) -> Fut,
-    Fut: Future<Output = Result<i64, RuntimeError>>,
-{
-    let rank = spec.shape.len();
-    let full_shape: Vec<usize> = if spec.dims == 1 {
-        vec![total_len_from_shape(spec.shape)]
-    } else if rank < spec.dims {
-        let mut s = spec.shape.to_vec();
-        s.resize(spec.dims, 1);
-        s
-    } else {
-        spec.shape.to_vec()
-    };
-
-    let mut selectors: Vec<ExprSel> = Vec::with_capacity(spec.dims);
-    let mut num_iter = 0usize;
-    let mut rp_iter = 0usize;
-    for d in 0..spec.dims {
-        let is_colon = (spec.colon_mask & (1u32 << d)) != 0;
-        let is_end = (spec.end_mask & (1u32 << d)) != 0;
-        if is_colon {
-            selectors.push(ExprSel::Colon);
-        } else if is_end {
-            selectors.push(ExprSel::Scalar(*full_shape.get(d).unwrap_or(&1)));
-        } else if let Some(pos) = spec.range_dims.iter().position(|&rd| rd == d) {
-            let (raw_st, raw_sp) = spec.range_params[rp_iter];
-            let dim_len = *full_shape.get(d).unwrap_or(&1);
-            let st = if let Some(expr) = &spec.range_start_exprs[rp_iter] {
-                resolve_end(dim_len, expr).await? as f64
-            } else {
-                raw_st
-            };
-            let sp = if let Some(expr) = &spec.range_step_exprs[rp_iter] {
-                resolve_end(dim_len, expr).await? as f64
-            } else {
-                raw_sp
-            };
-            rp_iter += 1;
-            let off = spec.range_end_exprs[pos].clone();
-            selectors.push(ExprSel::Range {
-                start: st as i64,
-                step: if sp >= 0.0 {
-                    sp as i64
-                } else {
-                    -(sp.abs() as i64)
-                },
-                end_off: off,
-            });
-        } else {
-            let v = spec.numeric.get(num_iter).ok_or_else(|| {
-                crate::interpreter::errors::mex("MissingNumericIndex", "missing numeric index")
-            })?;
-            num_iter += 1;
-            if let Some(idx) = index_scalar_from_value(v).await? {
-                if idx < 1 {
-                    return Err(crate::interpreter::errors::mex(
-                        "IndexOutOfBounds",
-                        "Index out of bounds",
-                    ));
-                }
-                selectors.push(ExprSel::Scalar(idx as usize));
-            } else {
-                match v {
-                    Value::Tensor(idx_t) => {
-                        let dim_len = *full_shape.get(d).unwrap_or(&1);
-                        let len = idx_t.shape.iter().product::<usize>();
-                        if len == dim_len {
-                            let mut vv = Vec::new();
-                            for (i, &val) in idx_t.data.iter().enumerate() {
-                                if val != 0.0 {
-                                    vv.push(i + 1);
-                                }
-                            }
-                            selectors.push(ExprSel::Indices(vv));
-                        } else {
-                            let mut vv = Vec::with_capacity(len);
-                            for &val in &idx_t.data {
-                                let idx = val as isize;
-                                if idx < 1 {
-                                    return Err(crate::interpreter::errors::mex(
-                                        "IndexOutOfBounds",
-                                        "Index out of bounds",
-                                    ));
-                                }
-                                vv.push(idx as usize);
-                            }
-                            selectors.push(ExprSel::Indices(vv));
-                        }
-                    }
-                    _ => {
-                        return Err(crate::interpreter::errors::mex(
-                            "UnsupportedIndexType",
-                            "Unsupported index type",
-                        ))
-                    }
-                }
-            }
-        }
-    }
-
-    let mut per_dim_indices: Vec<Vec<usize>> = Vec::with_capacity(spec.dims);
-    let mut selection_lengths: Vec<usize> = Vec::with_capacity(spec.dims);
-    let mut scalar_mask: Vec<bool> = Vec::with_capacity(spec.dims);
-    for (d, sel) in selectors.iter().enumerate().take(spec.dims) {
-        let dim_len = full_shape[d] as i64;
-        let idxs: Vec<usize> = match sel {
-            ExprSel::Colon => (1..=full_shape[d]).collect(),
-            ExprSel::Scalar(i) => vec![*i],
-            ExprSel::Indices(v) => v.clone(),
-            ExprSel::Range {
-                start,
-                step,
-                end_off,
-            } => {
-                let mut v = Vec::new();
-                let mut cur = *start;
-                let stp = *step;
-                let end_i = resolve_end(dim_len as usize, end_off).await?;
-                if stp == 0 {
-                    return Err(crate::interpreter::errors::mex(
-                        "IndexStepZero",
-                        "Index step cannot be zero",
-                    ));
-                }
-                if stp > 0 {
-                    while cur <= end_i {
-                        if cur < 1 || cur > dim_len {
-                            break;
-                        }
-                        v.push(cur as usize);
-                        cur += stp;
-                    }
-                } else {
-                    while cur >= end_i {
-                        if cur < 1 || cur > dim_len {
-                            break;
-                        }
-                        v.push(cur as usize);
-                        cur += stp;
-                    }
-                }
-                v
-            }
-        };
-        if idxs.iter().any(|&i| i == 0 || i > full_shape[d]) {
-            return Err(crate::interpreter::errors::mex(
-                "IndexOutOfBounds",
-                "Index out of bounds",
-            ));
-        }
-        selection_lengths.push(idxs.len());
-        per_dim_indices.push(idxs);
-        scalar_mask.push(matches!(sel, ExprSel::Scalar(_)));
-    }
-
-    let mut strides: Vec<usize> = vec![0; spec.dims];
-    let mut acc = 1usize;
-    for (d, stride) in strides.iter_mut().enumerate().take(spec.dims) {
-        *stride = acc;
-        acc *= full_shape[d];
-    }
-    let total_out: usize = per_dim_indices.iter().map(|v| v.len()).product();
-    if total_out == 0 {
-        let output_shape = if spec.dims == 1 {
-            vec![0, 1]
-        } else {
-            let mut dims_out: Vec<(usize, usize, bool)> = selection_lengths
-                .iter()
-                .enumerate()
-                .map(|(d, &len)| (d, len, scalar_mask.get(d).copied().unwrap_or(false)))
-                .collect();
-            while dims_out.len() > 2
-                && dims_out
-                    .last()
-                    .map(|&(_, len, is_scalar)| len == 1 && is_scalar)
-                    .unwrap_or(false)
-            {
-                dims_out.pop();
-            }
-            if dims_out.is_empty() {
-                vec![1, 1]
-            } else if dims_out.len() == 1 {
-                let (dim, len, _) = dims_out[0];
-                if dim == 1 {
-                    vec![1, len]
-                } else {
-                    vec![len, 1]
-                }
-            } else {
-                dims_out.into_iter().map(|(_, len, _)| len).collect()
-            }
-        };
-        return Ok(SlicePlan {
-            indices: Vec::new(),
-            output_shape,
-            selection_lengths,
-            dims: spec.dims,
-        });
-    }
-
-    let mut indices: Vec<u32> = Vec::with_capacity(total_out);
-    let mut idx = vec![0usize; spec.dims];
-    loop {
-        let mut lin = 0usize;
-        for d in 0..spec.dims {
-            let i0 = per_dim_indices[d][idx[d]] - 1;
-            lin += i0 * strides[d];
-        }
-        indices.push(lin as u32);
-        let mut d = 0usize;
-        while d < spec.dims {
-            idx[d] += 1;
-            if idx[d] < per_dim_indices[d].len() {
-                break;
-            }
-            idx[d] = 0;
-            d += 1;
-        }
-        if d == spec.dims {
-            break;
-        }
-    }
-
-    let output_shape = if spec.dims == 1 {
-        if total_out <= 1 {
-            vec![1, 1]
-        } else {
-            vec![total_out, 1]
-        }
-    } else {
-        let mut dims_out: Vec<(usize, usize, bool)> = selection_lengths
-            .iter()
-            .enumerate()
-            .map(|(d, &len)| (d, len, scalar_mask.get(d).copied().unwrap_or(false)))
-            .collect();
-        while dims_out.len() > 2
-            && dims_out
-                .last()
-                .map(|&(_, len, is_scalar)| len == 1 && is_scalar)
-                .unwrap_or(false)
-        {
-            dims_out.pop();
-        }
-        if dims_out.is_empty() {
-            vec![1, 1]
-        } else if dims_out.len() == 1 {
-            let (dim, len, _) = dims_out[0];
-            if dim == 1 {
-                vec![1, len]
-            } else {
-                vec![len, 1]
-            }
-        } else {
-            dims_out.into_iter().map(|(_, len, _)| len).collect()
-        }
-    };
-    Ok(SlicePlan {
-        indices,
-        output_shape,
-        selection_lengths,
-        dims: spec.dims,
-    })
 }
