+++ crates/runmat-vm/src/indexing/selectors.rs
@@ -1,3 +1,4 @@
+use crate::indexing::plan::total_len_from_shape;
 use crate::interpreter::errors::mex;
 use runmat_builtins::Value;
 use runmat_runtime::{
@@ -15,46 +16,6 @@ pub enum SliceSelector {
         values: Vec<usize>,
         output_shape: Vec<usize>,
     },
-}
-
-#[derive(Debug, Clone)]
-pub struct SlicePlan {
-    pub indices: Vec<u32>,
-    pub output_shape: Vec<usize>,
-    pub selection_lengths: Vec<usize>,
-    pub dims: usize,
-}
-
-fn cartesian_product<F: FnMut(&[usize])>(lists: &[Vec<usize>], mut f: F) {
-    let dims = lists.len();
-    if dims == 0 {
-        return;
-    }
-    let mut idx = vec![0usize; dims];
-    loop {
-        let current: Vec<usize> = (0..dims).map(|d| lists[d][idx[d]]).collect();
-        f(&current);
-        let mut d = 0usize;
-        while d < dims {
-            idx[d] += 1;
-            if idx[d] < lists[d].len() {
-                break;
-            }
-            idx[d] = 0;
-            d += 1;
-        }
-        if d == dims {
-            break;
-        }
-    }
-}
-
-fn total_len_from_shape(shape: &[usize]) -> usize {
-    if is_scalar_shape(shape) {
-        1
-    } else {
-        shape.iter().copied().product()
-    }
 }
 
 fn index_scalar_from_host_value(value: &Value) -> Option<i64> {
@@ -275,123 +236,4 @@ pub async fn build_slice_selectors(
         selectors.push(selector_from_value_dim(value, dim_len).await?);
     }
     Ok(selectors)
-}
-
-fn matlab_squeezed_shape(selection_lengths: &[usize], scalar_mask: &[bool]) -> Vec<usize> {
-    let mut dims: Vec<(usize, usize, bool)> = selection_lengths
-        .iter()
-        .enumerate()
-        .map(|(d, &len)| (d, len, scalar_mask.get(d).copied().unwrap_or(false)))
-        .collect();
-    while dims.len() > 2
-        && dims
-            .last()
-            .map(|&(_, len, is_scalar)| len == 1 && is_scalar)
-            .unwrap_or(false)
-    {
-        dims.pop();
-    }
-    let out: Vec<usize> = dims.into_iter().map(|(_, len, _)| len).collect();
-    if out.is_empty() {
-        vec![1, 1]
-    } else {
-        out
-    }
-}
-
-pub fn build_slice_plan(
-    selectors: &[SliceSelector],
-    dims: usize,
-    base_shape: &[usize],
-) -> VmResult<SlicePlan> {
-    let total_len = total_len_from_shape(base_shape);
-    if dims == 1 {
-        let list = selectors
-            .first()
-            .cloned()
-            .unwrap_or(SliceSelector::Indices(Vec::new()));
-        let indices = match &list {
-            SliceSelector::Colon => (1..=total_len).collect::<Vec<usize>>(),
-            SliceSelector::Scalar(i) => vec![*i],
-            SliceSelector::Indices(v) => v.clone(),
-            SliceSelector::LinearIndices { values, .. } => values.clone(),
-        };
-        if indices.iter().any(|&i| i == 0 || i > total_len) {
-            return Err(mex("IndexOutOfBounds", "Index out of bounds"));
-        }
-        let zero_based: Vec<u32> = indices.iter().map(|&i| (i - 1) as u32).collect();
-        let count = zero_based.len();
-        let shape = match list {
-            SliceSelector::LinearIndices { output_shape, .. } => output_shape,
-            _ if count <= 1 => vec![1, 1],
-            _ => vec![count, 1],
-        };
-        return Ok(SlicePlan {
-            indices: zero_based,
-            output_shape: shape,
-            selection_lengths: vec![count],
-            dims,
-        });
-    }
-
-    let mut selection_lengths = Vec::with_capacity(dims);
-    let mut per_dim_lists: Vec<Vec<usize>> = Vec::with_capacity(dims);
-    let mut scalar_mask: Vec<bool> = Vec::with_capacity(dims);
-    for (d, sel) in selectors.iter().enumerate().take(dims) {
-        let dim_len = base_shape.get(d).copied().unwrap_or(1);
-        let idxs = match sel {
-            SliceSelector::Colon => (1..=dim_len).collect::<Vec<usize>>(),
-            SliceSelector::Scalar(i) => vec![*i],
-            SliceSelector::Indices(v) => v.clone(),
-            SliceSelector::LinearIndices { values: v, .. } => v.clone(),
-        };
-        if idxs.iter().any(|&i| i == 0 || i > dim_len) {
-            return Err(mex("IndexOutOfBounds", "Index out of bounds"));
-        }
-        selection_lengths.push(idxs.len());
-        per_dim_lists.push(idxs);
-        scalar_mask.push(matches!(sel, SliceSelector::Scalar(_)));
-    }
-
-    let mut out_shape = matlab_squeezed_shape(&selection_lengths, &scalar_mask);
-    if selection_lengths.contains(&0) {
-        let selection_lengths = out_shape.clone();
-        return Ok(SlicePlan {
-            indices: Vec::new(),
-            output_shape: out_shape,
-            selection_lengths,
-            dims,
-        });
-    }
-
-    let mut base_norm = base_shape.to_vec();
-    if base_norm.len() < dims {
-        base_norm.resize(dims, 1);
-    }
-    let mut strides = vec![1usize; dims];
-    for d in 1..dims {
-        strides[d] = strides[d - 1] * base_norm[d - 1].max(1);
-    }
-
-    let mut indices = Vec::new();
-    cartesian_product(&per_dim_lists, |multi| {
-        let mut lin = 0usize;
-        for d in 0..dims {
-            let idx = multi[d] - 1;
-            lin += idx * strides[d];
-        }
-        indices.push(lin as u32);
-    });
-
-    let total_out: usize = selection_lengths.iter().product();
-    if total_out == 1 {
-        out_shape = vec![1, 1];
-    }
-    let selection_lengths = out_shape.clone();
-    Ok(SlicePlan {
-        indices,
-        output_shape: out_shape,
-        selection_lengths,
-        dims,
-    })
 }
